Fixed Response Errors

+++ src/main/java/com/example/reporting_api/model/response/Response.java
@@ -2,16 +2,14 @@ package com.example.reporting_api.model.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Response {
-    private Integer count;
-    private Integer total;
+    private Long count;
+    private Long total;
     private String currency;
